+++ src/game/CharacterHandler.cpp
@@ -1130,7 +1130,7 @@ void WorldSession::HandleAlterAppearanceOpcode(WorldPacket& recv_data)
         SendPacket(&data);
     }
 
-    _player->ModifyMoney(-int32(Cost));                     // it isn't free
+    _player->ModifyMoney(-int64(Cost));                     // it isn't free
     _player->GetAchievementMgr().UpdateAchievementCriteria(ACHIEVEMENT_CRITERIA_TYPE_GOLD_SPENT_AT_BARBER, Cost);
 
     _player->SetByteValue(PLAYER_BYTES, 2, uint8(bs_hair->hair_id));
